+++ packages/analysis/src/jsts/rules/S6767/decorator-indirect-prop-usage.ts
@@ -18,7 +18,12 @@
 import type { Rule, Scope, SourceCode } from 'eslint';
 import type estree from 'estree';
 import { getNodeParent } from '../helpers/ancestor.js';
-import { getVariableFromName, isFunctionNode, isIdentifier } from '../helpers/ast.js';
+import {
+  collectReferences,
+  getVariableFromName,
+  isFunctionNode,
+  isIdentifier,
+} from '../helpers/ast.js';
 import { isRequiredParserServices } from '../helpers/parser-services.js';
 import {
   getComponentIdentifier,
@@ -26,10 +31,9 @@ import {
 } from '../helpers/react/component-analysis.js';
 import { areSameTypeDeclarations, getTypeFromTreeNode } from '../helpers/type.js';
 import {
-  collectReferences,
   isNamedPropExpressionOrAlias,
   isWholePropsExpressionOrAlias,
-} from './prop-alias-resolution.js';
+} from './helpers/prop-alias-resolution.js';
 import { isSupportedWholePropsUsage } from './whole-props-usage.js';
 
 /**
